Interpreter

+++ compiler/src/main/kotlin/ast/ast.kt
@@ -334,9 +334,9 @@ class StateDecl(
     ctx: SourceContext,
     var ident: String,
     val multiStateCount: Int,
-    var red: Short,
-    var blue: Short,
-    var green: Short,
+    var red: Int,
+    var blue: Int,
+    var green: Int,
     val body: CodeBlock
 ) : Decl(ctx)
